Fixed terraform and scripts

+++ database/create-database.sql
@@ -4,6 +4,6 @@ USE projeto_pi;
 CREATE TABLE acumul_acelerate(
     acumul BIGINT,
     iterador INT,
-    time_spent DECIMAL(11, 9),
-    memory_usage DECIMAL(10,5)
+    time_spent DECIMAL(12, 9),
+    memory_usage INT
 );
